+++ src/snowflake/snowpark/session.py
@@ -4112,6 +4112,7 @@ class Session:
         *args: Any,
         statement_params: Optional[Dict[str, Any]] = None,
         log_on_exception: bool = False,
+        is_return_table: Optional[bool] = None,
         _emit_ast: bool = True,
     ) -> Any:
         """Calls a stored procedure by name.
@@ -4122,6 +4123,8 @@ class Session:
             statement_params: Dictionary of statement level parameters to be set while executing this action.
             log_on_exception: Log warnings if they arise when trying to determine if the stored procedure
                 as a table return type.
+            is_return_table: When set to a non-null value, it signifies whether the return type of sproc_name
+                is a table return type. This skips infer check and returns a dataframe with appropriate sql call.
 
         Example::
 
@@ -4161,6 +4164,7 @@ class Session:
             *args,
             statement_params=statement_params,
             log_on_exception=log_on_exception,
+            is_return_table=is_return_table,
             _emit_ast=_emit_ast,
         )